Completed length-based TLV encoding (BOLT-specific)

+++ strict_encoding/derive/tests/simple.rs
@@ -1,3 +1,17 @@
+// LNP/BP client-side-validation foundation libraries implementing LNPBP
+// specifications & standards (LNPBP-4, 7, 8, 9, 42, 81)
+//
+// Written in 2019-2021 by
+//     Dr. Maxim Orlovsky <orlovsky@pandoracore.com>
+//
+// To the extent possible under law, the author(s) have dedicated all
+// copyright and related and neighboring rights to this software to
+// the public domain worldwide. This software is distributed without
+// any warranty.
+//
+// You should have received a copy of the Apache 2.0 License along with this
+// software. If not, see <https://opensource.org/licenses/Apache-2.0>.
+
 #![allow(dead_code)]
 
 #[macro_use]
@@ -8,19 +22,19 @@ use std::collections::BTreeMap;
 #[derive(StrictEncode, StrictDecode)]
 struct Me(u8);
 
-#[derive(StrictEncode, StrictDecode)]
-#[strict_encoding(use_tlv)]
+#[derive(NetworkEncode, NetworkDecode)]
+#[network_encoding(use_tlv)]
 struct One {
     field_a: Vec<u8>,
 
-    #[strict_encoding(tlv = 1)]
+    #[network_encoding(tlv = 1)]
     tlv_int: Option<u16>,
 
-    #[strict_encoding(tlv = 2)]
+    #[network_encoding(tlv = 2)]
     tlv_int2: Option<u16>,
 
-    #[strict_encoding(unknown_tlvs)]
-    rest_of_tlvs: BTreeMap<u16, Box<[u8]>>,
+    #[network_encoding(unknown_tlvs)]
+    rest_of_tlvs: BTreeMap<usize, Box<[u8]>>,
 }
 
 #[derive(StrictEncode, StrictDecode)]
